requirements.txt
Flask>=1.1.1
requests>=2.20.0
numpy>=1.16.0
Pillow>=5.4.1
slackclient==1.3.0
